Make CValueLayout public
