+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/abi/x64/sysv/TypeClass.java
@@ -24,11 +24,11 @@
  */
 package jdk.internal.foreign.abi.x64.sysv;
 
+import jdk.incubator.foreign.CLinker;
 import jdk.incubator.foreign.GroupLayout;
 import jdk.incubator.foreign.MemoryLayout;
 import jdk.incubator.foreign.SequenceLayout;
 import jdk.incubator.foreign.ValueLayout;
-import jdk.internal.foreign.CValueLayout;
 import jdk.internal.foreign.Utils;
 
 import java.util.ArrayList;
@@ -107,11 +107,11 @@ class TypeClass {
     }
 
     private static ArgumentClassImpl argumentClassFor(MemoryLayout layout) {
-        if (!(layout instanceof CValueLayout)) {
+        if (!(layout instanceof CLinker.CValueLayout)) {
             throw new IllegalStateException("Unexpected value layout: could not determine ABI class");
         }
 
-        return switch (((CValueLayout) layout).kind()) {
+        return switch (((CLinker.CValueLayout) layout).kind()) {
             case CHAR, SHORT, INT, LONG, LONGLONG -> ArgumentClassImpl.INTEGER;
             case FLOAT, DOUBLE -> ArgumentClassImpl.SSE;
             case LONGDOUBLE -> ArgumentClassImpl.X87;
